Move all tests in a single `tests` module (#39)
